dags/bookstore_etl_dag_integrated.py
"""
INTEGRATED AIRFLOW DAG
Save this as: dags/bookstore_etl_dag_integrated.py
"""

from datetime import datetime, timedelta
from airflow import DAG
from airflow.operators.python import PythonOperator
from airflow.utils.dates import days_ago
import pandas as pd
import psycopg2
# ... (rest of imports)

# Configuration
DW_CONFIG = {
    'host': os.getenv('DW_POSTGRES_HOST', 'postgres-dw'),
    'port': os.getenv('DW_POSTGRES_PORT', '5432'),
    'database': os.getenv('DW_POSTGRES_DB', 'bookstore_dw'),
    'user': os.getenv('DW_POSTGRES_USER', 'dwuser'),
    'password': os.getenv('DW_POSTGRES_PASSWORD', 'dwpass123')
}

# Task functions
def extract_oltp_to_staging(**context):
    # Extract CSV to PostgreSQL staging
    pass

def transform_load_dimensions(**context):
    # Load dimension tables
    pass

def enrich_books_with_api(**context):
    # API enrichment
    pass

def load_fact_table(**context):
    # Load fact table
    pass

def export_ml_dataset(**context):
    # Export ML dataset
    pass

def data_quality_check(**context):
    # Quality check
    pass

# DAG definition
dag = DAG(
    'bookstore_integrated_etl',
    default_args=default_args,
    description='Integrated ETL Pipeline',
    schedule_interval='@daily',
    start_date=days_ago(1),
    catchup=False,
)

# Tasks
task1 = PythonOperator(task_id='extract', python_callable=extract_oltp_to_staging, dag=dag)
task2 = PythonOperator(task_id='transform', python_callable=transform_load_dimensions, dag=dag)
# ... etc

# Dependencies
task1 >> task2 >> task3 >> task4 >> task5 >> task6