requirements.txt
pandas==2.1.4
numpy==1.26.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
tqdm==4.66.1
python-dotenv==1.0.0
streamlit==1.31.0